UBUNTU: SAUCE: acpi battery -- delay first lookup of the battery until first use
When instantiating the battery object on to the acpi bus in the kernel
we talk to the BIOS to get the current battery state.  This can take
a long time and holds the acpi bus object locked for the duration.
This leads to any other object wishing to add itself that bus blocking.
This leads to unpredicatable delays of up to .3s when initialising the
hpet during boot depending on execution order.  It does not appear to be
necessary to inspect the actual battery until we first want the data.
Mark the battery as requiring update and continue.

Signed-off-by: Andy Whitcroft <apw@canonical.com>